feat: 입출력 기능 추가

+++ src/main/java/lotto/generator/NumberGenerator.java
@@ -9,5 +9,5 @@ import java.util.*;
 public interface NumberGenerator {
     Lotto generateLotto();
     Answer generateAnswer();
-    Lottos generateLottos(int lottoCount);
+    Lottos generateLottos(long lottoCount);
 }
